+++ src/main/java/com/iexec/core/pubsub/NotificationService.java
@@ -25,7 +25,7 @@ import org.springframework.stereotype.Service;
 @Service
 public class NotificationService {
 
-    private SimpMessagingTemplate sender;
+    private final SimpMessagingTemplate sender;
 
     public NotificationService(SimpMessagingTemplate sender) {
         this.sender = sender;
@@ -36,19 +36,5 @@ public class NotificationService {
         log.info("Sent TaskNotification [chainTaskId:{}, type:{}, workers:{}]",
                 taskNotification.getChainTaskId(), taskNotification.getTaskNotificationType(), taskNotification.getWorkersAddress());
     }
-
-    /* Test PubSub method
-    @Scheduled(fixedRate = 3000)
-    public void run(){
-        log.info("Check if results need to be uploaded");
-
-        //List<ReplicatesList> uploadableReplicates = taskService.getUploadableReplicates();
-        UploadResultMessage uploadResultMessage = UploadResultMessage.builder()
-                .taskId(null)
-                .workerAddress(null)
-                .build();
-        sendTaskNotification(uploadResultMessage);
-
-    }*/
 
 }