+++ src/main/java/com/iexec/core/replicate/ReplicateUpdatedEvent.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2023 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,18 +17,13 @@
 package com.iexec.core.replicate;
 
 import com.iexec.common.replicate.ReplicateStatusUpdate;
+import lombok.Builder;
+import lombok.Value;
 
-import lombok.*;
-
-@Data
+@Value
 @Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
 public class ReplicateUpdatedEvent {
-
-    private String chainTaskId;
-    private String walletAddress;
-    private ReplicateStatusUpdate replicateStatusUpdate;
+    String chainTaskId;
+    String walletAddress;
+    ReplicateStatusUpdate replicateStatusUpdate;
 }