+++ src/main/java/com/iexec/core/task/event/ResultUploadTimeoutEvent.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2023 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,15 +16,9 @@
 
 package com.iexec.core.task.event;
 
-import lombok.*;
+import lombok.Value;
 
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@Builder
+@Value
 public class ResultUploadTimeoutEvent {
-
-    private String chainTaskId;
+    String chainTaskId;
 }