+++ src/test/java/com/iexec/core/replicate/ReplicateServiceTests.java
@@ -1335,13 +1335,17 @@ class ReplicateServiceTests {
         when(replicatesRepository.findByChainTaskId(CHAIN_TASK_ID)).thenReturn(Optional.of(replicatesList));
         when(iexecHubService.repeatIsRevealedTrue(CHAIN_TASK_ID, WALLET_WORKER_1))
                 .thenReturn(true);
-        when(iexecHubService.getTaskDescriptionFromChain(CHAIN_TASK_ID)).thenReturn(Optional.of(task));
+        when(iexecHubService.getTaskDescription(CHAIN_TASK_ID)).thenReturn(task);
         when(resultService.isResultUploaded(CHAIN_TASK_ID)).thenReturn(false);
 
         assertThat(replicatesService.canUpdateReplicateStatus(CHAIN_TASK_ID, WALLET_WORKER_1, statusUpdate, null))
                 .isEqualTo(ReplicateStatusUpdateError.GENERIC_CANT_UPDATE);
-    }
 
+        verify(replicatesRepository).findByChainTaskId(CHAIN_TASK_ID);
+        verify(iexecHubService).repeatIsRevealedTrue(CHAIN_TASK_ID, WALLET_WORKER_1);
+        verify(iexecHubService).getTaskDescription(CHAIN_TASK_ID);
+        verify(resultService).isResultUploaded(CHAIN_TASK_ID);
+    }
 
     @Test
     void shouldNotAuthorizeUpdateOnContributeAndFinalizeDoneWhenTaskNotCompleted() {